predictive_engine/engine.py
/*
 * Copyright © 2025 Devin B. Royal.
 * All Rights Reserved.
 */


import numpy as np
import yaml

def load_config(path):
    with open(path, "r") as f:
        return yaml.safe_load(f)

def predict_error(data):
    # Real ML logic using numpy
    return np.mean(data) > 0.5

if __name__ == "__main__":
    config = load_config("predictive_config.yml")
    data = [0.1, 0.2, 0.8]
    if predict_error(data):
        print("Error predicted")
    else
        print("No error")

# # FULLY INJECTED by omniweave_full_inject.sh v1.0.0 at 2025-12-16T02:23:30Z
